backend/requirements.txt
absl-py==2.1.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
beautifulsoup4==4.12.3
billiard==4.2.1
bs4==0.0.2
cachetools==5.5.2
celery==5.5.3
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
datasketch==1.6.5
debugpy==1.8.5
decorator==5.1.1
executing==2.1.0
fastapi==0.116.1
Flask==2.2.5
Flask-Cors==3.0.9
flatbuffers==24.12.23
fonttools==4.51.0
gast==0.6.0
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-pasta==0.2.0
googleapis-common-protos==1.69.2
graphviz==0.20.3
greenlet==3.2.3
grpcio==1.68.1
h11==0.16.0
h5py==3.12.1
httplib2==0.22.0
httptools==0.6.4
idna==3.10
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.27.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.0.3
jinja2-time==0.2.0
joblib==1.3.2
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.7.0
kiwisolver==1.4.5
kombu==5.5.4
libclang==18.1.1
lxml==6.0.0
make==0.1.6.post2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
nest-asyncio==1.6.0
nltk==3.8.1
numpy==1.26.4
opt_einsum==3.4.0
optree==0.13.1
packaging==24.0
pandas==2.2.1
parso==0.8.4
pillow==10.3.0
platformdirs==4.3.2
pluggy==1.6.0
prompt_toolkit==3.0.47
proto-plus==1.26.1
protobuf==5.29.2
psutil==6.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.3.0
pydantic_core==2.33.2
Pygments==2.18.0
PyMuPDF==1.26.4
pyparsing==3.1.2
pytesseract==0.3.13
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2024.1
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
redis==6.3.0
regex==2023.12.25
requests==2.32.3
rich==13.9.4
rsa==4.9
scipy==1.13.0
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.42
sqlmodel==0.0.24
stack-data==0.6.3
starlette==0.47.2
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
text-unidecode==1.3
tornado==6.4.1
tqdm==4.66.2
traitlets==5.14.3
types-python-dateutil==2.9.0.20250708
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
unittest-xml-reporting==3.2.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.35.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0
